backend/requirements.txt
fastapi
uvicorn
python-dotenv
llama-index
llama-index-llms-google-genai
llama-index-embeddings-huggingface
llama-index-vector-stores-faiss
faiss-cpu
pypdf
python-multipart